+++ challenge/packages/api-graphql-with-dynamodb/src/modules/Videos/index.ts
@@ -1,4 +1,0 @@
-import './VideoTC';
-import './resolvers/Mutation.videosContest';
-import './resolvers/Query.videosByRanking';
-import './resolvers/Query.videosToCompare';