+++ challenge/packages/db-postgresdb/src/db.ts
@@ -1,5 +1,0 @@
-import { initialize } from '@ttoss/postgresdb';
-
-import * as models from './models';
-
-export const db = await initialize({ models });