+++ terezinha-farm/api/src/modules/FarmNotification/FarmNotificationTC.ts
@@ -1,9 +1,0 @@
-import { schemaComposer } from '@ttoss/graphql-api';
-
-export const FarmNotificationTC = schemaComposer.createObjectTC({
-  name: 'FarmNotification',
-  fields: {
-    farmId: 'ID!',
-    message: 'String!',
-  },
-});