+++ terezinha-farm/postgresdb/jest.config.ts
@@ -1,3 +1,0 @@
-import { jestRootConfig } from '@ttoss/config';
-
-export default jestRootConfig();