+++ terezinha-farm/postgresdb/src/models/index.ts
@@ -1,2 +1,0 @@
-export { Farm } from './Farm';
-export { User } from './User';