+++ terezinha-farm/vite-app/src/feature-flags.d.ts
@@ -1,1 +1,0 @@
-type FeatureFlags = 'nice-hi';